Update grammar and models to fix failing tests
All tests now pass!

The old filter_expr rules failed with a reduce-reduce conflict in the default LALR(1) mode due to a bug in jison: https://github.com/zaach/jison/issues/205

    $ jison src/xpath.jison src/xpath.jisonlex
    ...
    States with conflicts:
    State 3
      expr -> base_expr . #lookaheads= EOF OR...
      filter_expr -> base_expr . #lookaheads= ...

Aside: the old grammar (with filter_expr uncommented) did work in LR(1) mode, but it takes a LOOOOOONG time to generate which is annoying.

    $ jison -p lr src/xpath.jison src/xpath.jisonlex